requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
google-api-python-client==1.6.4
gTTS==1.2.2
gTTS-token==1.1.1
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
python-dotenv==0.7.1
requests==2.18.4
rsa==3.4.2
six==1.11.0
SpeechRecognition==3.8.1
uritemplate==3.0.0
urllib3==1.22
